requirements.txt
alembic==0.9.6
authy==2.1.5
Flask-DotEnv==0.1.1
Flask-Login==0.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
phonenumbers==8.8.5
